# fintrack-backend/public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "FinTrack API documentation (using Supabase JWT Bearer tokens)",
    "version": "1.0.0",
    "title": "FinTrack API"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
  "schemes": ["http"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Use: Bearer <access_token>"
    }
  },
...
  "paths": {
...
    "/receipts": {
      "get":{ "tags":["Receipts"], "summary":"List receipts","security":[{ "BearerAuth":[] }],"responses":{ "200":{"description":"Receipts list"} } },
      "post":{ "tags":["Receipts"], "summary":"Upload receipt","consumes":["multipart/form-data"],"security":[{ "BearerAuth":[] }],"parameters":[ { "name":"file","in":"formData","type":"file","required":true } ],"responses":{ "201":{"description":"Uploaded"}, "422":{"description":"Validation error"} } }
    },
...
    "/groups": {
      "get":{ "tags":["Groups"], "summary":"List groups","security":[{ "BearerAuth":[] }],"responses":{ "200":{"description":"Groups list"} } },
      "post":{ "tags":["Groups"], "summary":"Create group","consumes":["multipart/form-data"],"security":[{ "BearerAuth":[] }],"parameters":[ { "name":"name","in":"formData","type":"string","required":true } ],"responses":{ "201":{"description":"Created"}, "422":{"description":"Validation error"} } }
    },
    "/groups/{group}": {
      "parameters":[ { "name":"group","in":"path","type":"integer","required":true } ],
      "get":{ "tags":["Groups"], "summary":"Get single group","security":[{ "BearerAuth":[] }],"responses":{ "200":{"description":"Group"}, "404":{"description":"Not found"} } },
      "delete":{ "tags":["Groups"], "summary":"Delete group","security":[{ "BearerAuth":[] }], "responses":{ "204":{"description":"Deleted"} } }
    },
...
  }
}